requirements.txt
torch==2.3.0
torchvision==0.18.0
timm==1.0.3
datasets==2.19.1
huggingface_hub==0.23.0
xgboost==2.0.3
scikit-learn==1.4.2
scikit-image==0.23.2
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
requests==2.32.2
python-dotenv==1.0.1
pytest==8.2.0
pytest-cov==5.0.0
httpx==0.27.0
ruff==0.4.4
joblib==1.4.2
